KW-497/refactor: 로그아웃 로직을 LogoutButton 컴포넌트 내부로 이동
